Wallet improvements

+++ submodules/ton/tonlib-src/tdutils/td/utils/tl_parsers.h
@@ -168,8 +168,9 @@ class TlParser {
       data += sizeof(int32);
     } else {
       check_len(sizeof(int32));
-      result_len = data[1] + (data[2] << 8) + (data[3] << 16) + (data[4] << 24) + (static_cast<uint64>(data[5]) << 32) +
-                   (static_cast<uint64>(data[6]) << 40) + (static_cast<uint64>(data[7]) << 48);
+      result_len = narrow_cast<size_t>(data[1] + (data[2] << 8) + (data[3] << 16) + (data[4] << 24) +
+                                       (static_cast<uint64>(data[5]) << 32) + (static_cast<uint64>(data[6]) << 40) +
+                                       (static_cast<uint64>(data[7]) << 48));
       if (result_len > std::numeric_limits<size_t>::max() - 3) {
         set_error("Too big string found");
         return T();
